Add files via upload
city bike which launched in 2013 is based on two categories, subscribers and non - subscribers.  it is the nations largest bike share program operating out of New York City. this bike share system with 27,000 bikes and over 1700 stations across Manhattan, Brooklyn, Queens, the Bronx, Jersey City, and Hoboken, has become an essential part of a transportation network for New Yorkers at any day and time of the year.